Borrado Logico Post y regrese /api/ en server.use
